# docs/api/openapi.yml
components:
  schemas:
...
    DashboardtypesDashboard:
      properties:
        createdAt:
          format: date-time
          type: string
        createdBy:
          type: string
        data:
          $ref: '#/components/schemas/DashboardtypesStorableDashboardData'
        id:
          type: string
        locked:
          type: boolean
        org_id:
          type: string
        updatedAt:
          format: date-time
          type: string
        updatedBy:
          type: string
      type: object
    DashboardtypesGettablePublicDasbhboard:
      properties:
        defaultTimeRange:
          type: string
        publicPath:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    DashboardtypesGettablePublicDashboardData:
      properties:
        dashboard:
          $ref: '#/components/schemas/DashboardtypesDashboard'
        publicDashboard:
          $ref: '#/components/schemas/DashboardtypesGettablePublicDasbhboard'
      type: object
    DashboardtypesPostablePublicDashboard:
      properties:
        defaultTimeRange:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    DashboardtypesStorableDashboardData:
      additionalProperties: {}
      type: object
    DashboardtypesUpdatablePublicDashboard:
      properties:
        defaultTimeRange:
          type: string
        timeRangeEnabled:
          type: boolean
      type: object
    ErrorsJSON:
      properties:
        code:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorsResponseerroradditional'
          type: array
        message:
          type: string
        url:
          type: string
      required:
      - code
      - message
      type: object
    ErrorsResponseerroradditional:
      properties:
        message:
          type: string
      type: object
...
    PromotetypesPromotePath:
      properties:
        indexes:
          items:
            $ref: '#/components/schemas/PromotetypesWrappedIndex'
          type: array
        path:
          type: string
        promote:
          type: boolean
      type: object
    PromotetypesWrappedIndex:
      properties:
        column_type:
          type: string
        granularity:
          type: integer
        type:
          type: string
      type: object
    Querybuildertypesv5AggregationBucket:
      properties:
        alias:
          type: string
        anomalyScores:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeries'
          type: array
        index:
          type: integer
        lowerBoundSeries:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeries'
          type: array
        meta:
          properties:
            unit:
              type: string
          type: object
        predictedSeries:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeries'
          type: array
        series:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeries'
          nullable: true
          type: array
        upperBoundSeries:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeries'
          type: array
      type: object
    Querybuildertypesv5Bucket:
      properties:
        step:
          format: double
          type: number
      type: object
...
    Querybuildertypesv5ColumnDescriptor:
      properties:
        aggregationIndex:
          format: int64
          type: integer
        columnType:
          $ref: '#/components/schemas/Querybuildertypesv5ColumnType'
        description:
          type: string
        fieldContext:
          $ref: '#/components/schemas/TelemetrytypesFieldContext'
        fieldDataType:
          $ref: '#/components/schemas/TelemetrytypesFieldDataType'
        meta:
          properties:
            unit:
              type: string
          type: object
        name:
          type: string
        queryName:
          type: string
        signal:
          $ref: '#/components/schemas/TelemetrytypesSignal'
        unit:
          type: string
      required:
      - name
      type: object
    Querybuildertypesv5ColumnType:
      enum:
      - group
      - aggregation
      type: string
...
    Querybuildertypesv5ExecStats:
      description: Execution statistics for the query, including rows scanned, bytes
        scanned, and duration.
      properties:
        bytesScanned:
          minimum: 0
          type: integer
        durationMs:
          minimum: 0
          type: integer
        rowsScanned:
          minimum: 0
          type: integer
        stepIntervals:
          additionalProperties:
            minimum: 0
            type: integer
          type: object
      type: object
...
    Querybuildertypesv5Label:
      properties:
        key:
          $ref: '#/components/schemas/TelemetrytypesTelemetryFieldKey'
        value: {}
      type: object
...
    Querybuildertypesv5QueryData:
      oneOf:
      - $ref: '#/components/schemas/Querybuildertypesv5TimeSeriesData'
      - $ref: '#/components/schemas/Querybuildertypesv5ScalarData'
      - $ref: '#/components/schemas/Querybuildertypesv5RawData'
      properties:
        results:
          items: {}
          nullable: true
          type: array
      type: object
...
    Querybuildertypesv5QueryRangeResponse:
      description: 'Response from the v5 query range endpoint. The data.results array
        contains typed results depending on the requestType: TimeSeriesData for time_series,
        ScalarData for scalar, or RawData for raw requests.'
      properties:
        data:
          $ref: '#/components/schemas/Querybuildertypesv5QueryData'
        meta:
          $ref: '#/components/schemas/Querybuildertypesv5ExecStats'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5RequestType'
        warning:
          $ref: '#/components/schemas/Querybuildertypesv5QueryWarnData'
      type: object
...
    Querybuildertypesv5QueryWarnData:
      properties:
        message:
          type: string
        url:
          type: string
        warnings:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5QueryWarnDataAdditional'
          type: array
      type: object
    Querybuildertypesv5QueryWarnDataAdditional:
      properties:
        message:
          type: string
      type: object
    Querybuildertypesv5RawData:
      properties:
        nextCursor:
          type: string
        queryName:
          type: string
        rows:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5RawRow'
          nullable: true
          type: array
      type: object
    Querybuildertypesv5RawRow:
      properties:
        data:
          additionalProperties: {}
          nullable: true
          type: object
        timestamp:
          format: date-time
          type: string
      type: object
...
    Querybuildertypesv5RequestType:
      enum:
      - scalar
      - time_series
      - raw
      - raw_stream
      - trace
      type: string
    Querybuildertypesv5ScalarData:
      properties:
        columns:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5ColumnDescriptor'
          nullable: true
          type: array
        data:
          items:
            items: {}
            type: array
          nullable: true
          type: array
        queryName:
          type: string
      type: object
...
    Querybuildertypesv5TimeSeries:
      properties:
        labels:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5Label'
          type: array
        values:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TimeSeriesValue'
          nullable: true
          type: array
      type: object
    Querybuildertypesv5TimeSeriesData:
      properties:
        aggregations:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5AggregationBucket'
          nullable: true
          type: array
        queryName:
          type: string
      type: object
    Querybuildertypesv5TimeSeriesValue:
      properties:
        bucket:
          $ref: '#/components/schemas/Querybuildertypesv5Bucket'
        partial:
          type: boolean
        timestamp:
          format: int64
          type: integer
        value:
          format: double
          type: number
        values:
          items:
            format: double
            type: number
          type: array
      type: object
...
    RenderErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorsJSON'
        status:
          type: string
      required:
      - status
      - error
      type: object
...
    TelemetrytypesFieldContext:
      enum:
      - metric
      - log
      - span
      - resource
      - attribute
      - body
      type: string
    TelemetrytypesFieldDataType:
      enum:
      - string
      - bool
      - float64
      - int64
      - number
      type: string
...
    TelemetrytypesSignal:
      enum:
      - traces
      - logs
      - metrics
      type: string
...
    TelemetrytypesTelemetryFieldKey:
      properties:
        description:
          type: string
        fieldContext:
          $ref: '#/components/schemas/TelemetrytypesFieldContext'
        fieldDataType:
          $ref: '#/components/schemas/TelemetrytypesFieldDataType'
        name:
          type: string
        signal:
          $ref: '#/components/schemas/TelemetrytypesSignal'
        unit:
          type: string
      required:
      - name
      type: object
...
    TypesIdentifiable:
      properties:
        id:
          type: string
      required:
      - id
      type: object
...
  securitySchemes:
    api_key:
      description: API Keys
      in: header
      name: Hanzo O11y-Api-Key
      type: apiKey
    tokenizer:
      bearerFormat: Tokenizer
      description: Tokens generated by the tokenizer
      scheme: bearer
      type: http
info:
  description: OpenTelemetry-Native Logs, Metrics and Traces in a single pane
  title: Hanzo O11y
  version: ""
openapi: 3.0.3
paths:
...
  /api/v1/dashboards/{id}/public:
    delete:
      deprecated: false
      description: This endpoint deletes the public sharing config and disables the
        public sharing of a dashboard
      operationId: DeletePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Delete public dashboard
      tags:
      - dashboard
    get:
      deprecated: false
      description: This endpoint returns public sharing config for a dashboard
      operationId: GetPublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/DashboardtypesGettablePublicDasbhboard'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Get public dashboard
      tags:
      - dashboard
    post:
      deprecated: false
      description: This endpoint creates public sharing config and enables public
        sharing of the dashboard
      operationId: CreatePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardtypesPostablePublicDashboard'
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/TypesIdentifiable'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: Created
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Create public dashboard
      tags:
      - dashboard
    put:
      deprecated: false
      description: This endpoint updates the public sharing config for a dashboard
      operationId: UpdatePublicDashboard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardtypesUpdatablePublicDashboard'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Update public dashboard
      tags:
      - dashboard
...
  /api/v1/logs/promote_paths:
    get:
      deprecated: false
      description: This endpoints promotes and indexes paths
      operationId: ListPromotedAndIndexedPaths
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: '#/components/schemas/PromotetypesPromotePath'
                    nullable: true
                    type: array
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - VIEWER
      - tokenizer:
        - VIEWER
      summary: Promote and index paths
      tags:
      - logs
    post:
      deprecated: false
      description: This endpoints promotes and indexes paths
      operationId: HandlePromoteAndIndexPaths
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/PromotetypesPromotePath'
              nullable: true
              type: array
      responses:
        "201":
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - EDITOR
      - tokenizer:
        - EDITOR
      summary: Promote and index paths
      tags:
      - logs
...
  /api/v1/public/dashboards/{id}:
    get:
      deprecated: false
      description: This endpoint returns the sanitized dashboard data for public access
      operationId: GetPublicDashboardData
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/DashboardtypesGettablePublicDashboardData'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - anonymous:
        - public-dashboard:read
      summary: Get public dashboard data
      tags:
      - dashboard
  /api/v1/public/dashboards/{id}/widgets/{idx}/query_range:
    get:
      deprecated: false
      description: This endpoint return query range results for a widget of public
        dashboard
      operationId: GetPublicDashboardWidgetQueryRange
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: idx
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Querybuildertypesv5QueryRangeResponse'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - anonymous:
        - public-dashboard:read
      summary: Get query range result
      tags:
      - dashboard
...